+++ lib/src/player.native.dart
@@ -20,6 +20,10 @@ class VideoControllerImplementation extends VideoController {
   int? get subId => _subId;
   int? _subId;
 
+  // The rotation of the video in 90 degree increments.
+  int get rotation => _rotation;
+  int _rotation = 0;
+
   StreamSubscription? _eventSubscription;
   String? _source;
   var _seeking = false;
@@ -81,8 +85,11 @@ class VideoControllerImplementation extends VideoController {
                     loading.value) {
                   final width = e['width'] as double;
                   final height = e['height'] as double;
-                  if (width != videoSize.value.width ||
+                  final rotation = e['rotation'] as int? ?? 0;
+                  if (_rotation != rotation ||
+                      width != videoSize.value.width ||
                       height != videoSize.value.height) {
+                    _rotation = rotation;
                     videoSize.value = width > 0 && height > 0
                         ? Size(width, height)
                         : Size.zero;
@@ -506,6 +513,7 @@ class VideoControllerImplementation extends VideoController {
   }
 
   void _close() {
+    _rotation = 0;
     _seeking = false;
     mediaInfo.value = null;
     videoSize.value = Size.zero;